+++ REXCompiler/omptests/rex-simd/benchmark/axpy/rose/rose_axpy_float_unroll_sve.c
@@ -80,7 +80,7 @@ void axpy(float *X,float *Y,float a)
     svfloat32_t __vec38 = svmul_f32_m(__pg0,__vec37,__vec36);
     svfloat32_t __vec39 = svadd_f32_m(__pg0,__vec38,__vec35);
     svst1(__pg0,&Y[i + svcntw() * 7],__vec39);
-    __pg0 = svwhilelt_b32(i,(unsigned long )102399999);
+    __pg0 = svwhilelt_b32((unsigned long )i,(unsigned long )102399999);
   }
 }
 // Debug functions
