+++ REXCompiler/omptests/rex-simd/benchmark/spmv/rose/rose_spmv_float_sve.c
@@ -114,7 +114,7 @@ int main(int argc,char *argv[])
       svfloat32_t __vec3 = svmul_f32_m(__pg0,__vec2,__vec1);
       svfloat32_t __vec4 = svadd_f32_m(__pg0,__vec3,__part0);
       __part0 = (__vec4);
-      __pg0 = svwhilelt_b32(idx,(unsigned long )(ia[row + 1] - 1));
+      __pg0 = svwhilelt_b32((unsigned long )idx,(unsigned long )(ia[row + 1] - 1));
     }
     __pg0 = svptrue_b32();
     sum += svaddv(__pg0,__part0);
